refactor: splitted cardano-service in several modules and some renamings
